Rename `orm_factory` to `orm_utils`
This module now does more than just construct new ORM classes and the
`_utils` suffix, while terrible, fits the pattern used elsewhere.
