+++ tests/acceptance/comparative_booster_study/test_variables_lib.py
@@ -5,7 +5,7 @@ import pytest
 import sqlalchemy.orm
 
 from databuilder.ehrql import Dataset
-from databuilder.orm_factory import orm_class_from_ql_table
+from databuilder.orm_utils import orm_class_from_ql_table
 from databuilder.query_language import EventFrame, Series, table
 
 from .variables_lib import create_sequential_variables
