+++ tests/spec/tables.py
@@ -3,7 +3,7 @@ import datetime
 import sqlalchemy.orm
 
 from databuilder.codes import SNOMEDCTCode
-from databuilder.orm_factory import orm_class_from_ql_table
+from databuilder.orm_utils import orm_class_from_ql_table
 from databuilder.tables import EventFrame, PatientFrame, Series, table
